% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{my.file.browse}
\alias{my.file.browse}
\title{Text-based interactive file selection.}
\usage{
my.file.browse(root = getwd(), multiple = F)
}
\arguments{
\item{root}{the root directory to explore
(default current working directory)}

\item{multiple}{boolean specifying whether to allow
multiple files to be selected}
}
\value{
character vector of selected files.
}
\description{
Text-based interactive file selection.
}
\examples{
fileList <- my.file.browse()
}
